Allow qualifier to be modified prior to use
The plugin assumes that the qualifier will be stored in the form of subdomain.domain.tld in the tenant table. Some also just store the subdomain (as the domain will change depending on environment, dev, stage, live). This commit adds a config var to model (modifyQualifier) with the key 'method'. The only method supported now is str_replace and another config var model.modifyQualifier.remove gives the pattern which is removed completely.